Refactor Router & Layout. Delete content from Register/Login Pages to avoid conflicts
